Moved Intervals enum into statistics/lib
