+++ src/kpi/statistics/lib/index.ts
@@ -1,1 +1,2 @@
 export * from './transformMapToObject';
+export * from './intervals';
