+++ src/kpi/statistics/lib/intervals.ts
